реализовали кнопку удаления и очистки. осталось привести в порядок форму и добавить кнопку добавления и редактирования

+++ DBTest/DBWinform/Form1.cs
@@ -21,18 +21,28 @@ namespace DBWinform
         {
             InitializeComponent();
             dataGridView.SelectionMode = DataGridViewSelectionMode.FullRowSelect;
-            List<Users> users = loader.Load();
+
+            BindingList<Users> users = loader.Load();
             dataGridView.DataSource = users;
         }
 
         private void DeleteButton_Click(object sender, EventArgs e)
         {
-            DataGridViewRow row = dataGridView.SelectedRows[0];
-            Users user = row.DataBoundItem as Users;
-            loader.DeleteSelectUser(user.Login);
-
-
-
+            DialogResult result = MessageBox.Show("Вы действительно хотите удалить эту запись?", "Внимамние", MessageBoxButtons.YesNo);
+            if (result == DialogResult.Yes)
+            {
+                DataGridViewRow row = dataGridView.SelectedRows[0];
+                Users user = row.DataBoundItem as Users;
+                loader.DeleteSelectUser(user.Login);
+            }
+        }
+        private void clearButton_Click(object sender, EventArgs e)
+        {
+            DialogResult result = MessageBox.Show("Вы действительно хотите удалить все записи?", "Внимамние", MessageBoxButtons.YesNo);
+            if (result == DialogResult.Yes)
+            {
+                loader.ClearAllUsers();
+            }
         }
     }
 }
